+++ src/utils/index.js
@@ -1,4 +1,4 @@
 export * from './api';
 export * from './environment';
 export * from './misc';
-export * from './organisation';
+export * from './document';
